Move validation function within ModelAnalysis class
